LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/out-of-document-rule-set.https_include=CSPExemption-expected.txt

CONSOLE MESSAGE: Refused to load https://localhost:9443/resources/testdriver.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://localhost:9443/resources/testdriver-vendor.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://localhost:9443/resources/testdriver-actions.js because it does not appear in the script-src directive of the Content Security Policy.

FAIL It should accept the speculation rules when the main page has strict CSP. assert_in_array: Prefetch should work for request {"shouldPrefetch":true,"status":200,"redirect":false,"useRelativeUrlForCandidate":false,"useRelativeUrlForSpeculationRulesSet":false,"useUtf8EncodingForSpeculationRulesSet":true,"failCors":false,"useValidSpeculationRulesHeaderValue":true,"useInnerListInSpeculationRulesHeaderValue":false,"useEmptySpeculationRulesSet":false,"useValidJsonForSpeculationRulesSet":true,"useValidUrlForSpeculationRulesSet":true,"useValidMimeTypeForSpeculationRulesSet":true,"strictCSP":true}. value undefined not in array ["prefetch", "prefetch;anonymous-client-ip"]
